keeping updated

+++ src/process_bulk_images.py
@@ -29,15 +29,15 @@ def multiple_replace(dict, text):
     regex = re.compile("(%s)" % "|".join(map(re.escape, dict.keys())))
     return regex.sub(lambda mo: dict[mo.string[mo.start():mo.end()]], text)
 
-path_to_lookup = glob.glob(os.path.join(DEFAULT_PATH+"etl\\pictures", "*.jp*"))
+path_to_lookup = glob.glob(os.path.join(DEFAULT_PATH+"\\etl\\pictures", "*.*"))
 
 success_text = []
 errors = []
 
 for image in path_to_lookup:
     try:
-        os.rename(image.upper(), multiple_replace(dict, image.upper()))
-        image = multiple_replace(dict, image)
+        image = image.upper()
+        os.rename(image, multiple_replace(dict, image))
         basename = os.path.basename(image)
         (filename, _ext) = os.path.splitext(basename)
 
@@ -46,9 +46,11 @@ for image in path_to_lookup:
     except Exception as ex:
         errors.append(f"{filename}: {ex}\n")
     finally:
-        os.remove(image)
-       
-fs.create_file(file="bulk-image_ended.bool")
+        exists = os.path.isfile(image)
+        if exists:
+            os.remove(image)
+
+fs.create_file(file="bulk_image_ended.bool")
 
 if success_text:
     fs.write_nlines_file(file="ready4db.txt", text=success_text, method="a")
